Creation of high-level CoupledSolver class
