+++ coupling_components/coupled_solvers/iqni.py
@@ -1,12 +1,12 @@
 from coconut.tools import create_instance, pass_on_parameters
-from coconut.coupling_components.coupled_solvers.gauss_seidel import CoupledSolverGaussSeidel
+from coconut.coupling_components.coupled_solvers.coupled_solver import CoupledSolver
 
 
 def create(parameters):
     return CoupledSolverIQNI(parameters)
 
 
-class CoupledSolverIQNI(CoupledSolverGaussSeidel):
+class CoupledSolverIQNI(CoupledSolver):
     def __init__(self, parameters):
         super().__init__(parameters)
